PGPE: Add Multicasting Enable/Disable Build Flag
Original-Change-Id: Id55e4ec522f35071c3661d26a96a802365355cc1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84261
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>

+++ import/chips/p10/common/ppe/powmanlib/ppehw_common.h
@@ -195,7 +195,7 @@ enum PPE_MULTICAST_TYPES
 
 // use this to override undesired queued putscom with nop
 #define PPE_PUTSCOM_NOQ(addr, data)                               \
-    putscom_norc(addr, data);
+    putscom_norc((uint32_t)addr, data);
 
 #define PPE_PUTSCOM(addr, data)                                   \
     putscom_norc(PPE_QUEUED_SCOM(addr), data);
